+++ RELEASENOTES.md
@@ -249,6 +249,7 @@
         Material3-themed Composables (PlayPauseButton, NextButton,
         PreviousButton, SeekBackButton, SeekForwardButton, RepeatButton,
         ShuffleButton, MuteButton) to it.
+    *   Add support for placing a media route button in the `PlayerView`.
 *   Downloads:
 *   OkHttp extension:
 *   Cronet extension: